switch from double to float

+++ src/geometry/geometry.hpp
@@ -19,7 +19,7 @@ public:
 
     bool renderOn = true;
     bool isLight = false;
-    double brightness;
+    float brightness;
 
     void setMaterial(Material & m);
     virtual vec4 getNormal(vec4 & pos, ray & inRay) = 0;
@@ -31,7 +31,7 @@ public:
     }
 
     void translate(vec4 & v);
-    void rotate(double angle, vec4 & axis);
+    void rotate(float angle, vec4 & axis);
     void scale(vec4 & v);
 
     virtual Hit trace(ray & inRay) = 0;
